KiriKini/requirements.txt
boto3==1.10.8
botocore==1.13.8
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-sslserver==0.21
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
facebook-sdk==3.1.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
